requirements.txt
discord.py==2.3.1
gTTS==2.3.2
psycopg2==2.9.6
PyNaCl==1.5.0
python-dotenv==1.0.0